+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-08-14 18:08:16
-// Commit: fix(network): fix memory leak by cancelling network subscription
+// Updated at: 2026-08-14 19:51:57
+// Commit: test(network): add unit tests for network business logic
 
 class networkComponentState {
-  final String commitRef = '3617';
-  final String lastMessage = 'fix(network): fix memory leak by cancelling network subscription';
-  final DateTime timestamp = DateTime.parse('2026-08-14T18:08:16+05:00');
+  final String commitRef = '3619';
+  final String lastMessage = 'test(network): add unit tests for network business logic';
+  final DateTime timestamp = DateTime.parse('2026-08-14T19:51:57+05:00');
 }